Support PHP 8.0; Drop 5.6, 7.0 & 7.1
Fixes #88

To support PHP 8.0 we have to update PHPUnit: only versions 8 and 9
support PHP 8.0, see https://phpunit.de/supported-versions.html

Then we have to update the test suite as required by breaking changes in
PHPUnit 8.  And this makes the test suite to stop working with older
versions of PHP: 5.6, 7.0 and 7.1, so we drop support for these.

Anyway these old versions of PHP reached EOL a while ago, see
https://en.wikipedia.org/wiki/PHP#Release_history